requirements.txt
google-api-python-client
google-cloud-storage
google-cloud-secret-manager
google-auth